% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polyarun.R
\name{polyaurn}
\alias{polyaurn}
\title{Polyaurn function}
\usage{
polyaurn(red, black, trials)
}
\arguments{
\item{red}{number of red balls}

\item{black}{number of black balls}

\item{trials}{number of trials in game}
}
\description{
Another implementation of the urn game from Games of Life by Karl Sigmund (ISBN: 978-0198547839)
This time, an example of runaway drift with an unlikely twist.
}
\examples{
urn(1,1,1000)
}
\keyword{gamesoflife,}
\keyword{genetics}
\keyword{polyaurn,}
