% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_drift.R
\name{random_drift}
\alias{random_drift}
\title{Random drift function}
\usage{
random_drift(red, black, trials, show.black = FALSE)
}
\arguments{
\item{red}{number of red balls}

\item{black}{number of black balls}

\item{trials}{number of trials in game}
}
\description{
Two coloured ball game with entangled probabilities.
}
\examples{
random_drift(red = 10000, black = 10000, trials = 100)
}
\keyword{gamesoflife,}
\keyword{genetics}
\keyword{urn,}
