% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drift_many_colours.R
\name{random_drifts}
\alias{random_drifts}
\title{Drift with many competitors function}
\usage{
random_drifts(colours = 5, trials, size, rel.freq = NULL)
}
\arguments{
\item{colours}{number of different colours of balls}

\item{trials}{the number of trials the game runs for}

\item{size}{size of the initial population of balls}

\item{rel.freq}{proportions}
}
\description{
What happens if we have a population of coloured balls and let them drift?
}
\examples{
random_drifts(colours = 10, trials = 1000, size = 1000)
}
\keyword{gamesoflife,}
\keyword{genetics}
\keyword{urn,}
