% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection_drifts.R
\name{selection_drifts}
\alias{selection_drifts}
\title{Drift with many competitors, adjust selection for one lucky colour function}
\usage{
selection_drifts(colours = 5, trials, size, rel.freq = NULL,
  s = 1.04)
}
\arguments{
\item{colours}{number of different colours of balls}

\item{trials}{the number of trials the game runs for}

\item{size}{size of the initial population of balls}

\item{rel.freq}{starting frequencies for the colours. If NULL, frequencies are drawn from the uniform distribution.}

\item{s}{selection coefficient for the first colour}
}
\description{
Change the selection coefficient, s, to adjust how many copies of that colour are left
}
\examples{
selection_drifts(colours = 10, trials = 1000, size = 1000)
}
\keyword{drift}
\keyword{gamesoflife,}
\keyword{selction,}
\keyword{urn,}
