% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Urn.R
\name{urn}
\alias{urn}
\title{Urn function}
\usage{
urn(red, black, trials)
}
\arguments{
\item{red}{number of red balls}

\item{black}{number of black balls}

\item{trials}{number of trials in game}
}
\description{
Implementation of the urn game.
}
\examples{
urn(1,1,1000)
}
\keyword{gamesoflife,}
\keyword{genetics}
\keyword{urn,}
