% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Urn.R
\name{urn}
\alias{urn}
\title{Urn function}
\usage{
urn(red, black, trials)
}
\arguments{
\item{number}{of red/black balls and trial number}
}
\description{
Implementation of the urn game from Games of Life by Karl Sigmund (ISBN: 978-0198547839)
}
\examples{
urn(1,1,1000)
}
\keyword{gamesoflife,}
\keyword{genetics}
\keyword{urn,}
